Rename .gitattributes to readme
A scoreboard for live premier league games and custom retro game scoreboards. It can be set to show all premier league games scores for the day or it can be set to show only your teams score. There is also a custom match feature that allows the display of custom matches with all information entered into the custom_matches.json file.